bootstrap jumbotron
hello guys,
in this file, i made a jumbotron with navbar on top and in body part i made a 3 columns for 3 different paragraphs. and at the end of the page i given a footer. you can use this code if u thinking it is helpful.
and change accordingly.
thanks 
|<>|